server/requirements.txt
Flask==3.1.0
wikipedia
googletrans==4.0.2
google-generativeai
google-genai
python-dotenv
gtts
pydub
ffmpeg-python
mutagen
mongoengine
asyncio
aiohttp
boto3
flask_cors
recharts
PyPDF2
python-docx